wip: style edge js refactoring
