+++ elixir/web-ng/assets/js/lib/god_view/rendering_style_edge_topology_methods.js
@@ -1,0 +1,32 @@
+export const godViewRenderingStyleEdgeTopologyMethods = {
+  edgeTopologyClassFromLabel(label) {
+    const text = String(label == null ? "" : label).trim().toUpperCase()
+    if (text.includes(" ENDPOINT ")) return "endpoints"
+    if (text.includes(" INFERRED ")) return "inferred"
+    return "backbone"
+  },
+  edgeTopologyClass(edge) {
+    const explicit = String(edge?.topologyClass || "").trim().toLowerCase()
+    if (explicit === "inferred" || explicit === "endpoints" || explicit === "backbone") {
+      return explicit
+    }
+    return this.edgeTopologyClassFromLabel(edge?.label || "")
+  },
+  edgeEnabledByTopologyLayer(edge) {
+    const classCounts = edge?.topologyClassCounts
+    if (classCounts && typeof classCounts === "object") {
+      const showBackbone =
+        Number(classCounts.backbone || 0) > 0 && this.topologyLayers.backbone !== false
+      const showInferred =
+        Number(classCounts.inferred || 0) > 0 && this.topologyLayers.inferred === true
+      const showEndpoints =
+        Number(classCounts.endpoints || 0) > 0 && this.topologyLayers.endpoints === true
+      return showBackbone || showInferred || showEndpoints
+    }
+
+    const topologyClass = this.edgeTopologyClass(edge)
+    if (topologyClass === "inferred") return this.topologyLayers.inferred === true
+    if (topologyClass === "endpoints") return this.topologyLayers.endpoints === true
+    return this.topologyLayers.backbone !== false
+  },
+}
